Extract wallet context to its own file
