+++ src/pages/Wallet/WalletHomePage.tsx
@@ -21,7 +21,7 @@ import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import { AnimatePresence, motion, useViewportScroll } from 'framer-motion'
 import { Lock, QrCode, RefreshCw, Send } from 'lucide-react'
-import { useCallback, useContext, useEffect, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import styled from 'styled-components'
 
@@ -32,13 +32,13 @@ import { FloatingPanel, Section } from '../../components/PageComponents/PageCont
 import Spinner from '../../components/Spinner'
 import TransactionItem from '../../components/TransactionItem'
 import { useGlobalContext } from '../../contexts/global'
+import { SimpleTx, useWalletContext } from '../../contexts/wallet'
 import { ReactComponent as AlephiumLogoSVG } from '../../images/alephium_logo_monochrome.svg'
 import { appHeaderHeight, deviceBreakPoints } from '../../style/globalStyles'
 import { getHumanReadableError } from '../../utils/api'
 import { useInterval } from '../../utils/hooks'
 import { abbreviateAmount, calAmountDelta } from '../../utils/numbers'
 import { loadStoredSettings } from '../../utils/settings'
-import { SimpleTx, WalletContext } from './WalletRootPage'
 
 dayjs.extend(relativeTime)
 
@@ -46,7 +46,7 @@ const WalletHomePage = () => {
   const history = useHistory()
   const { wallet, setSnackbarMessage, client, lockWallet, currentUsername, currentNetwork } = useGlobalContext()
   const [balance, setBalance] = useState<bigint | undefined>(undefined)
-  const { networkPendingTxLists, loadedTxList, setLoadedTxList } = useContext(WalletContext)
+  const { networkPendingTxLists, loadedTxList, setLoadedTxList } = useWalletContext()
   const [totalNumberOfTx, setTotalNumberOfTx] = useState(0)
   const [isLoading, setIsLoading] = useState(false)
   const [isHeaderCompact, setIsHeaderCompact] = useState(false)
